@@ -5,16 +5,17 @@ import (
 	"crypto/rsa"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
+	"log"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"strings"
 	"testing"
-	"log"
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/dgrijalva/jwt-go"
+	"github.com/form3tech-oss/jwt-go"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -27,16 +28,18 @@ func fatal(err error) {
 }
 
 func (m *MockClient) Get(url string) (*http.Response, error) {
-	if url != "https://www.googleapis.com/oauth2/v1/certs" {
+	if url != "https://qgraph.us.auth0.com/.well-known/jwks.json" {
 		return &http.Response{
 			StatusCode: 404,
-			Body:       ioutil.NopCloser(bytes.NewReader([]byte("Page not found"))),
+			Body:       io.NopCloser(bytes.NewReader([]byte("Page not found"))),
 		}, nil
 	}
-	cert, err := ioutil.ReadFile(pubKeyPath)
+	cert, err := os.ReadFile(pubKeyPath)
 	fatal(err)
-	json := fmt.Sprintf(`{"default":"%s"}`, strings.ReplaceAll(string(cert), "\n", `\n`))
-	r := ioutil.NopCloser(bytes.NewReader([]byte(json)))
+	sub_cert := strings.Replace(string(cert), "-----BEGIN CERTIFICATE-----", "", 1)
+	sub_cert = strings.Replace(sub_cert, "-----END CERTIFICATE-----", "", 1)
+	json := fmt.Sprintf(`{"keys":[{"kid":"default","x5c":["%s"]}]}`, strings.ReplaceAll(sub_cert, "\n", ""))
+	r := io.NopCloser(bytes.NewReader([]byte(json)))
 	return &http.Response{
 		StatusCode: 200,
 		Body:       r,
@@ -51,7 +54,7 @@ func performRequest(r http.Handler, method, path string, jwt *string, body *stri
 		req, _ = http.NewRequest(method, path, strings.NewReader(*body))
 	}
 	if jwt != nil {
-		req.Header.Add("Authorization", "Bearer "+ *jwt)
+		req.Header.Add("Authorization", "Bearer "+*jwt)
 	}
 
 	w := httptest.NewRecorder()
@@ -65,7 +68,6 @@ var (
 )
 
 var (
-	verifyKey    *rsa.PublicKey
 	signKey      *rsa.PrivateKey
 	router       *gin.Engine
 	signedString string
@@ -74,20 +76,17 @@ var (
 func init() {
 	Client = &MockClient{}
 
-	signBytes, err := ioutil.ReadFile(privKeyPath)
+	signBytes, err := os.ReadFile(privKeyPath)
 	fatal(err)
 	signKey, err = jwt.ParseRSAPrivateKeyFromPEM(signBytes)
 	fatal(err)
-	verifyBytes, err := ioutil.ReadFile(pubKeyPath)
-	fatal(err)
-	verifyKey, err = jwt.ParseRSAPublicKeyFromPEM(verifyBytes)
-	fatal(err)
 
+	gin.SetMode(gin.TestMode)
 	// Grab our router
 	router = SetupRouter()
 
 	type MyCustomClaims struct {
-		Email string `json:"email,omitempty"`
+		Email string `json:"https://qgraph.org/email,omitempty"`
 		jwt.StandardClaims
 	}
 
@@ -95,8 +94,8 @@ func init() {
 	claims := MyCustomClaims{
 		"me@robokache.com",
 		jwt.StandardClaims{
-			Issuer:   "accounts.google.com",
-			Audience: "297705140796-41v2ra13t7mm8uvu2dp554ov1btt80dg.apps.googleusercontent.com",
+			Issuer:   "https://qgraph.us.auth0.com/",
+			Audience: []string{"https://qgraph.org/api"},
 		},
 	}
 
@@ -107,12 +106,11 @@ func init() {
 
 // Ensure that we don't have failing setup functions
 func TestSetupFunctions(t *testing.T) {
-  err := clearDB()
-  assert.Nil(t, err)
+	err := clearDB()
+	assert.Nil(t, err)
 
-  err = loadSampleData()
-  assert.Nil(t, err)
-
+	err = loadSampleData()
+	assert.Nil(t, err)
 }
 
 func TestGetWithMissingBearer(t *testing.T) {
@@ -134,13 +132,14 @@ func TestGetWithMalformedBearer(t *testing.T) {
 }
 
 func TestGetDocumentsNotLoggedIn(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	w := performRequest(router, "GET", "/api/document", nil, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response []map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 
 	// Should be able to see only public documents (3)
@@ -148,13 +147,14 @@ func TestGetDocumentsNotLoggedIn(t *testing.T) {
 }
 
 func TestGetDocuments(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	w := performRequest(router, "GET", "/api/document", &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response []map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 
 	// Should be able to see my documents (4) + you public documents (2)
@@ -162,14 +162,15 @@ func TestGetDocuments(t *testing.T) {
 }
 
 func TestGetDocumentsNoParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
-  // Gets root documents
+	// Gets root documents
 	w := performRequest(router, "GET", "/api/document?has_parent=false", &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response []map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 
 	// Should be able to see my root documents (2) + you root public documents (1)
@@ -180,14 +181,15 @@ func TestGetDocumentsNoParent(t *testing.T) {
 }
 
 func TestGetDocumentsHasParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
-    // Gets root documents
+	// Gets root documents
 	w := performRequest(router, "GET", "/api/document?has_parent=true", &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response []map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 
 	// Should be able to see my child documents (2) + you child public documents (1)
@@ -198,139 +200,151 @@ func TestGetDocumentsHasParent(t *testing.T) {
 }
 
 func TestGetMePrivateDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Can get my own private document
 	hashedID, _ := idToHash(0)
-	w := performRequest(router, "GET", "/api/document/" + hashedID, &signedString, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	// Check that the response looks ok
 	var response map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 	assert.Equal(t, hashedID, response["id"])
 	assert.Equal(t, true, response["owned"])
 }
 
 func TestGetYouPrivateDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	hashedID, err := idToHash(6)
 	assert.Nil(t, err)
-	w := performRequest(router, "GET", "/api/document/" + hashedID, &signedString, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, &signedString, nil)
+	fmt.Println(w.Code)
+	fmt.Println(http.StatusNotFound)
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
 func TestGetYouShareableDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Can get other's shareable documents
 	hashedID, err := idToHash(5)
 	assert.Nil(t, err)
-	w := performRequest(router, "GET", "/api/document/" + hashedID, &signedString, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response map[string]interface{}
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 	assert.Equal(t, hashedID, response["id"])
 	assert.Equal(t, false, response["owned"])
 }
 
 func TestGetPublicDocumentNotLoggedIn(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	hashedID, err := idToHash(3)
 	assert.Nil(t, err)
 	// Can get public document
-	w := performRequest(router, "GET", "/api/document/" + hashedID, nil, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, nil, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response map[string]interface{}
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 }
 
 func TestGetShareableDocumentNotLoggedIn(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	hashedID, err := idToHash(1)
 	assert.Nil(t, err)
 	// Can get shareable documents
-	w := performRequest(router, "GET", "/api/document/" + hashedID, nil, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, nil, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response map[string]interface{}
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 }
 
 func TestGetPrivateDocumentNotLoggedIn(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	hashedID, err := idToHash(0)
 	assert.Nil(t, err)
 	// Cannot get private documents
-	w := performRequest(router, "GET", "/api/document/" + hashedID, nil, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID, nil, nil)
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
 func TestGetChildren(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Can see all of my own child documents
 	hashedID, _ := idToHash(1)
-	w := performRequest(router, "GET", "/api/document/" + hashedID + "/children", &signedString, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID+"/children", &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 	var response []map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 	assert.Equal(t, 2, len(response))
 
 	// Can see public and shareable child documents if the document is not owned by me
 	hashedID, _ = idToHash(5)
 	assert.Nil(t, err)
-	w = performRequest(router, "GET", "/api/document/" + hashedID + "/children", &signedString, nil)
+	w = performRequest(router, "GET", "/api/document/"+hashedID+"/children", &signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 	assert.Equal(t, 2, len(response))
 }
 
 func TestGetChildrenPrivateDoc(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Can't view other users' child documents
 	hashedID, _ := idToHash(6)
-	w := performRequest(router, "GET", "/api/document/" + hashedID + "/children", &signedString, nil)
+	w := performRequest(router, "GET", "/api/document/"+hashedID+"/children", &signedString, nil)
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
 func TestGetPutData(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	id, _ := idToHash(1)
 	requestBody := "This is a string to test the data saving functionality"
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s/data`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s/data`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	w = performRequest(router, "GET",
-			fmt.Sprintf(`/api/document/%s/data`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s/data`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Equal(t, requestBody, w.Body.String())
 }
 
 func TestGetNoData(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	id, _ := idToHash(1)
 	w := performRequest(router, "GET",
-			fmt.Sprintf(`/api/document/%s/data`, id),
-			&signedString, nil)
+		fmt.Sprintf(`/api/document/%s/data`, id),
+		&signedString, nil)
 
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Equal(t, "", w.Body.String())
@@ -338,13 +352,14 @@ func TestGetNoData(t *testing.T) {
 
 // Check that we can post an empty document and it is valid
 func TestPostEmpty(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	requestBody := `{}`
 	w := performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusCreated, w.Code)
 	var response map[string]interface{}
-	err := json.Unmarshal([]byte(w.Body.String()), &response)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Nil(t, err)
 
 	newDocumentIDHash := response["id"].(string)
@@ -355,9 +370,9 @@ func TestPostEmpty(t *testing.T) {
 
 	// Check that the document was created with private visibility
 	w = performRequest(router, "GET",
-			fmt.Sprintf(`/api/document/%s`, newDocumentIDHash),
-			&signedString, &requestBody)
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+		fmt.Sprintf(`/api/document/%s`, newDocumentIDHash),
+		&signedString, &requestBody)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Nil(t, err)
 	// JSON response numbers are parsed as floats
@@ -366,28 +381,29 @@ func TestPostEmpty(t *testing.T) {
 
 // Check that we get 401 error on POST route when not logged in
 func TestPostNotLoggedIn(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	requestBody := `{ "visibility" : 4 }`
 	w := performRequest(router, "POST", "/api/document", nil, &requestBody)
 	assert.Equal(t, http.StatusUnauthorized, w.Code)
 }
 
-
 // Test the shortcut route to add a child with data
 func TestPostChildWithData(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	id, _ := idToHash(1)
 	requestBody := "This is a string to test the data saving functionality"
 	w := performRequest(router, "POST",
-			fmt.Sprintf(`/api/document/%s/children`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s/children`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	var response map[string]interface{}
-  err := json.Unmarshal([]byte(w.Body.String()), &response)
-  assert.Nil(t, err)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
+	assert.Nil(t, err)
 
 	newDocumentIDHash := response["id"].(string)
 	log.Println(newDocumentIDHash)
@@ -398,9 +414,9 @@ func TestPostChildWithData(t *testing.T) {
 
 	// Check that the document was created with the same visibility as parent
 	w = performRequest(router, "GET",
-			fmt.Sprintf(`/api/document/%s`, newDocumentIDHash),
-			&signedString, &requestBody)
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
+		fmt.Sprintf(`/api/document/%s`, newDocumentIDHash),
+		&signedString, &requestBody)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Nil(t, err)
 	// JSON response numbers are parsed as floats
@@ -408,21 +424,21 @@ func TestPostChildWithData(t *testing.T) {
 
 	// Check that the document data was saved correctly
 	w = performRequest(router, "GET",
-			fmt.Sprintf(`/api/document/%s/data`, newDocumentIDHash),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s/data`, newDocumentIDHash),
+		&signedString, &requestBody)
 	assert.Equal(t, requestBody, w.Body.String())
 }
 
-
 func TestPostDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 	requestBody := `{ "visibility" : 4 }`
 	w := performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusCreated, w.Code)
 
 	var response map[string]interface{}
-  err := json.Unmarshal([]byte(w.Body.String()), &response)
-  assert.Nil(t, err)
+	err := json.Unmarshal(w.Body.Bytes(), &response)
+	assert.Nil(t, err)
 
 	// Check that the ID was returned
 	createdID, err := hashToID(response["id"].(string))
@@ -431,49 +447,51 @@ func TestPostDocument(t *testing.T) {
 }
 
 func TestPostDocumentWithMetadata(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
-  metadata := map[string]interface{}{
-    "hasAnswers" : true,
-    "questionName" : "My cool question",
-  }
+	metadata := map[string]interface{}{
+		"hasAnswers":   true,
+		"questionName": "My cool question",
+	}
 
-  metadataSerialized, err := json.Marshal(metadata)
-  assert.Nil(t, err)
+	metadataSerialized, err := json.Marshal(metadata)
+	assert.Nil(t, err)
 
-  requestBody := fmt.Sprintf(`{ "metadata" : %s }`, metadataSerialized)
+	requestBody := fmt.Sprintf(`{ "metadata" : %s }`, metadataSerialized)
 	w := performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusCreated, w.Code)
 
 	var response map[string]interface{}
-  err = json.Unmarshal([]byte(w.Body.String()), &response)
-  assert.Nil(t, err)
+	err = json.Unmarshal(w.Body.Bytes(), &response)
+	assert.Nil(t, err)
 
 	// Check that the metadata exists on object
 	createdID := response["id"].(string)
-	w = performRequest(router, "GET", "/api/document/" + createdID, &signedString, nil)
-	assert.Nil(t, err)
+	w = performRequest(router, "GET", "/api/document/"+createdID, &signedString, nil)
 
-	err = json.Unmarshal([]byte(w.Body.String()), &response)
-  assert.NotNil(t, response["metadata"])
-  t.Log(response)
-  assert.Equal(t, metadata["hasAnswers"],
-                  response["metadata"].(map[string]interface{})["hasAnswers"])
+	err = json.Unmarshal(w.Body.Bytes(), &response)
+	assert.Nil(t, err)
+	assert.NotNil(t, response["metadata"])
+	t.Log(response)
+	assert.Equal(t, metadata["hasAnswers"],
+		response["metadata"].(map[string]interface{})["hasAnswers"])
 }
 
 func TestPostDocumentWithParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	parentID, _ := idToHash(1)
 	requestBody := fmt.Sprintf(
-			`{ "parent" : "%s", "visibility" : %d }`,
+		`{ "parent" : "%s", "visibility" : %d }`,
 		parentID, shareable)
 	w := performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusCreated, w.Code)
 
-  var response map[string]interface{}
-  err := json.Unmarshal([]byte(w.Body.String()), &response)
-  assert.Nil(t, err)
+	var response map[string]interface{}
+	err := json.Unmarshal(w.Body.Bytes(), &response)
+	assert.Nil(t, err)
 
 	// Check that the ID was returned
 	createdID, err := hashToID(response["id"].(string))
@@ -482,12 +500,13 @@ func TestPostDocumentWithParent(t *testing.T) {
 }
 
 func TestPostDocumentInvalidParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Parent document is not owned by me
 	parentID, _ := idToHash(4)
 	requestBody := fmt.Sprintf(
-			`{ "parent" : "%s", "visibility" : %d }`,
+		`{ "parent" : "%s", "visibility" : %d }`,
 		parentID, shareable)
 	w := performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusBadRequest, w.Code)
@@ -495,68 +514,73 @@ func TestPostDocumentInvalidParent(t *testing.T) {
 	// Parent document has less visibility
 	parentID, _ = idToHash(0)
 	requestBody = fmt.Sprintf(
-			`{ "parent" : "%s", "visibility" : %d }`,
+		`{ "parent" : "%s", "visibility" : %d }`,
 		parentID, shareable)
 	w = performRequest(router, "POST", "/api/document", &signedString, &requestBody)
 	assert.Equal(t, http.StatusBadRequest, w.Code)
 }
 
 func TestPutDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	requestBody := fmt.Sprintf(`{ "visibility" : %d }`, private)
 	idHash, _ := idToHash(1)
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, idHash),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, idHash),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusOK, w.Code)
 }
 
 func TestPutDocumentNotOwned(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	requestBody := fmt.Sprintf(`{ "visibility" : %d }`, private)
 	idHash, _ := idToHash(4)
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, idHash),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, idHash),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusForbidden, w.Code)
 }
 
 func TestPutDocumentNotVisible(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	requestBody := fmt.Sprintf(`{ "visibility" : %d }`, private)
 	idHash, _ := idToHash(6)
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, idHash),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, idHash),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusNotFound, w.Code)
 }
 
 func TestPutDocumentModifyParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	newParentID, _ := idToHash(0)
 	requestBody := fmt.Sprintf(`{ "parent" : "%s", "visibility" : %d }`,
-							   newParentID, private)
+		newParentID, private)
 	id, _ := idToHash(1)
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusOK, w.Code)
 }
 
 func TestPutDocumentInvalidParent(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	// Not enough visibility on parent
 	newParentID, _ := idToHash(0)
 	requestBody := fmt.Sprintf(`{ "parent" : "%s" }`, newParentID)
 	id, _ := idToHash(1)
 	w := performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusBadRequest, w.Code)
 
 	// Not owned by current user
@@ -564,8 +588,8 @@ func TestPutDocumentInvalidParent(t *testing.T) {
 	requestBody = fmt.Sprintf(`{ "parent" : "%s" }`, newParentID)
 	id, _ = idToHash(1)
 	w = performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusBadRequest, w.Code)
 
 	// Doesn't exist
@@ -573,27 +597,29 @@ func TestPutDocumentInvalidParent(t *testing.T) {
 	requestBody = fmt.Sprintf(`{ "parent" : "%s" }`, newParentID)
 	id, _ = idToHash(1)
 	w = performRequest(router, "PUT",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, &requestBody)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, &requestBody)
 	assert.Equal(t, http.StatusBadRequest, w.Code)
 }
 
 func TestDeleteDocument(t *testing.T) {
-	clearDB(); loadSampleData()
+	clearDB()
+	loadSampleData()
 
 	id, _ := idToHash(1)
 	w := performRequest(router, "DELETE",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, nil)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, nil)
 	assert.Equal(t, http.StatusOK, w.Code)
 
 	// Can't delete other user's document
 	id, _ = idToHash(4)
 	w = performRequest(router, "DELETE",
-			fmt.Sprintf(`/api/document/%s`, id),
-			&signedString, nil)
+		fmt.Sprintf(`/api/document/%s`, id),
+		&signedString, nil)
 	assert.Equal(t, http.StatusForbidden, w.Code)
 }
+
 // Benchmark to test how the application handles large files
 func BenchmarkGetPutLargeData(b *testing.B) {
 	var testBytes []byte
@@ -616,11 +642,11 @@ func BenchmarkGetPutLargeData(b *testing.B) {
 		b.StartTimer()
 
 		performRequest(router, "PUT",
-				fmt.Sprintf(`/api/document/%s/data`, id),
-				&signedString, &testString)
+			fmt.Sprintf(`/api/document/%s/data`, id),
+			&signedString, &testString)
 		performRequest(router, "GET",
-				fmt.Sprintf(`/api/document/%s/data`, id),
-				&signedString, nil)
+			fmt.Sprintf(`/api/document/%s/data`, id),
+			&signedString, nil)
 
 		b.StopTimer()
 		clearDB()
