@@ -1,9 +1,10 @@
 package robokache
 
 import (
+	"fmt"
 	"net/http"
 	"strings"
-	"fmt"
+
 	"github.com/gin-gonic/gin"
 	log "github.com/sirupsen/logrus"
 )
@@ -20,33 +21,40 @@ func init() {
 func handleErr(c *gin.Context, err error) {
 	errorMsg := err.Error()
 	errorResponse := map[string]string{
-		"message" : errorMsg,
+		"message": errorMsg,
 	}
-	if strings.HasPrefix(errorMsg, "Bad Request") {
+	if strings.HasPrefix(errorMsg, "bad request") {
 		c.JSON(400, errorResponse)
-	} else if strings.HasPrefix(errorMsg, "Unauthorized") {
+	} else if strings.HasPrefix(errorMsg, "unauthorized") {
 		c.JSON(401, errorResponse)
-	} else if strings.HasPrefix(errorMsg, "Forbidden") {
+	} else if strings.HasPrefix(errorMsg, "forbidden") {
 		c.JSON(403, errorResponse)
-	} else if strings.HasPrefix(errorMsg, "Not Found") {
+	} else if strings.HasPrefix(errorMsg, "not found") {
 		c.JSON(404, errorResponse)
 	} else {
-		log.WithFields(log.Fields{"error" : err}).
-		WithContext(c).
-		Error("Internal Server Error")
+		log.WithFields(log.Fields{"error": err}).
+			WithContext(c).
+			Error("Internal Server Error")
 		// Rewrite error message so that we don't expose it to the user
 		errorResponse["message"] = "Internal Server Error"
 		c.JSON(500, errorResponse)
 	}
 }
 
-
 // AddGUI adds the GUI endpoints
 func AddGUI(r *gin.Engine) {
 	// Serve HTML
 	r.LoadHTMLGlob("./web/index.html")
+	r.Static("/assets", "./web/assets")
 	r.GET("/", func(c *gin.Context) {
 		c.HTML(http.StatusOK, "index.html", nil)
+	})
+	r.GET("/auth_config.json", func(c *gin.Context) {
+		c.JSON(200, gin.H{
+			"domain":   "qgraph.us.auth0.com",
+			"clientId": "sgJrK1gGAbzrXwUp0WG7jAV0ivCIF6jr",
+			"audience": "https://qgraph.org/api",
+		})
 	})
 
 	// Serve static files (openapi.yml)
@@ -58,7 +66,7 @@ type GetDocumentQuery struct {
 	HasParent *bool `form:"has_parent"`
 }
 
-func GetUserEmail(c *gin.Context) (*string) {
+func GetUserEmail(c *gin.Context) *string {
 	val, ok := c.Get("userEmail")
 	if ok {
 		email, _ := val.(*string)
@@ -71,7 +79,8 @@ func GetUserEmail(c *gin.Context) (*string) {
 func SetupRouter() *gin.Engine {
 	r := gin.Default()
 
-	r.Use(AddUserToContext)
+	// validateUser also adds userEmail to Context
+	r.Use(validateUser)
 
 	api := r.Group("/api")
 
@@ -85,7 +94,7 @@ func SetupRouter() *gin.Engine {
 			var queryParams GetDocumentQuery
 			err := c.ShouldBindQuery(&queryParams)
 			if err != nil {
-				handleErr(c, fmt.Errorf("Bad Request: Error parsing query parameters"))
+				handleErr(c, fmt.Errorf("bad request: Error parsing query parameters"))
 				return
 			}
 
@@ -197,156 +206,157 @@ func SetupRouter() *gin.Engine {
 			c.JSON(http.StatusOK, documents)
 		})
 	}
+	{
+		api.POST("/document/:id/children", func(c *gin.Context) {
+			userEmail := GetUserEmail(c)
+			if userEmail == nil {
+				handleErr(c,
+					fmt.Errorf("unauthorized: You must be logged in to add a document"))
+				return
+			}
 
-	api.POST("/document/:id/children", func(c *gin.Context) {
-		userEmail := GetUserEmail(c)
-		if userEmail == nil {
-			handleErr(c,
-			fmt.Errorf("Unauthorized: You must be logged in to add a document."))
-			return
-		}
+			// Get document id
+			parentID, err := hashToID(c.Param("id"))
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Get document id
-		parentID, err := hashToID(c.Param("id"))
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Get the parent so we can set the default visibility
+			// to the visibility of the parent
+			parent, err := GetDocument(userEmail, parentID)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
+			newDoc := Document{
+				Parent:     &parent.ID,
+				Visibility: parent.Visibility,
+				Owner:      *userEmail,
+			}
 
-		// Get the parent so we can set the default visibility
-		// to the visibility of the parent
-		parent, err := GetDocument(userEmail, parentID)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
-		newDoc := Document{
-			Parent: &parent.ID,
-			Visibility: parent.Visibility,
-			Owner: *userEmail,
-		}
+			// Add the document to the database
+			newDocID, err := PostDocument(newDoc)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Add the document to the database
-		newDocID, err := PostDocument(newDoc)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Write data to disk
+			err = SetData(newDocID, c.Request.Body)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Write data to disk
-		err = SetData(newDocID, c.Request.Body)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Convert ID to hash
+			newDocIDHash, err := idToHash(newDocID)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Convert ID to hash
-		newDocIDHash, err := idToHash(newDocID)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Return
+			response := make(map[string]string)
+			response["id"] = newDocIDHash
+			c.JSON(http.StatusOK, response)
+		})
+		api.POST("/document", func(c *gin.Context) {
+			userEmail := GetUserEmail(c)
+			if userEmail == nil {
+				handleErr(c,
+					fmt.Errorf("unauthorized: You must be logged in to add a document"))
+				return
+			}
 
-		// Return
-		response := make(map[string]string)
-		response["id"] = newDocIDHash
-		c.JSON(http.StatusOK, response)
-	})
+			// Parse the document from JSON
+			doc := makeDefaultDoc()
+			err := c.ShouldBindJSON(&doc)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
+			// Set the document owner from the user's email
+			doc.Owner = *userEmail
 
-	api.POST("/document", func(c *gin.Context) {
-		userEmail := GetUserEmail(c)
-		if userEmail == nil {
-			handleErr(c,
-			fmt.Errorf("Unauthorized: You must be logged in to add a document."))
-			return
-		}
+			// Convert user given hashes to IDs
+			err = doc.addID()
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Parse the document from JSON
-		doc := makeDefaultDoc()
-		err := c.ShouldBindJSON(&doc)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
-		// Set the document owner from the user's Google Auth
-		doc.Owner = *userEmail
+			// Add document to DB
+			newID, err := PostDocument(doc)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Convert user given hashes to IDs
-		err = doc.addID()
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Convert new ID to hash
+			hashedID, err := idToHash(newID)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Add document to DB
-		newID, err := PostDocument(doc)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Return
+			response := make(map[string]string)
+			response["id"] = hashedID
+			c.JSON(http.StatusCreated, response)
+		})
+	}
+	{
+		api.PUT("/document/:id", func(c *gin.Context) {
+			userEmail := GetUserEmail(c)
+			if userEmail == nil {
+				handleErr(c,
+					fmt.Errorf("unauthorized: You must be logged in to edit a document"))
+				return
+			}
 
-		// Convert new ID to hash
-		hashedID, err := idToHash(newID)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Get document id
+			id, err := hashToID(c.Param("id"))
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Return
-		response := make(map[string]string)
-		response["id"] = hashedID
-		c.JSON(http.StatusCreated, response)
-	})
-	api.PUT("/document/:id", func(c *gin.Context) {
-		userEmail := GetUserEmail(c)
-		if userEmail == nil {
-			handleErr(c,
-			fmt.Errorf("Unauthorized: You must be logged in to edit a document."))
-			return
-		}
+			// Parse the document from JSON
+			var doc Document
+			err = c.ShouldBindJSON(&doc)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
+			// Convert user-given parent hash and ID hash to IDs
+			err = doc.addID()
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Get document id
-		id, err := hashToID(c.Param("id"))
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Check we have permission to update this document
+			existingDoc, err := GetDocumentForEditing(*userEmail, id)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Parse the document from JSON
-		var doc Document
-		err = c.ShouldBindJSON(&doc)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
-		// Convert user-given parent hash and ID hash to IDs
-		err = doc.addID()
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Set the document owner from the user's Google Auth
+			doc.Owner = *userEmail
+			// Set the document ID based on the URL param
+			doc.ID = id
 
-		// Check we have permission to update this document
-		existingDoc, err := GetDocumentForEditing(*userEmail, id)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
+			// Add document to DB
+			err = EditDocument(doc, existingDoc)
+			if err != nil {
+				handleErr(c, err)
+				return
+			}
 
-		// Set the document owner from the user's Google Auth
-		doc.Owner = *userEmail
-		// Set the document ID based on the URL param
-		doc.ID = id
-
-		// Add document to DB
-		err = EditDocument(doc, existingDoc)
-		if err != nil {
-			handleErr(c, err)
-			return
-		}
-
-		log.WithFields(
-			log.Fields{"doc" : fmt.Sprintf("%+v", doc)}).Debug("Updating document")
+			log.WithFields(
+				log.Fields{"doc": fmt.Sprintf("%+v", doc)}).Debug("Updating document")
 
 			response := make(map[string]string)
 			c.JSON(http.StatusOK, response)
@@ -355,7 +365,7 @@ func SetupRouter() *gin.Engine {
 			userEmail := GetUserEmail(c)
 			if userEmail == nil {
 				handleErr(c,
-				fmt.Errorf("Unauthorized: You must be logged in to edit a document."))
+					fmt.Errorf("unauthorized: You must be logged in to edit a document"))
 				return
 			}
 
@@ -383,11 +393,13 @@ func SetupRouter() *gin.Engine {
 			response := make(map[string]string)
 			c.JSON(http.StatusOK, response)
 		})
+	}
+	{
 		api.DELETE("/document/:id", func(c *gin.Context) {
 			userEmail := GetUserEmail(c)
 			if userEmail == nil {
 				handleErr(c,
-				fmt.Errorf("Unauthorized: You must be logged in to delete a document."))
+					fmt.Errorf("unauthorized: You must be logged in to delete a document"))
 				return
 			}
 
@@ -415,5 +427,5 @@ func SetupRouter() *gin.Engine {
 			c.JSON(http.StatusOK, response)
 		})
 		return r
-
+	}
 }
